refactor(client): extract shared useUrlParamSettings hook from AppChat/AppCanvas
Both pages implemented an almost byte-identical effect for reading
model/style/outfmt/temp/history/var_* query params, applying them, and
stripping them from the URL. The duplication had already drifted (AppChat
also stripped prefill/send; AppCanvas didn't apply parsed var_* values
anywhere). Centralizing it in one hook means future fixes only need to
land once.

Fixes #1752
